debian: add systemd service for EdgeRouter v2.0
  * check INIT env var for either "sysv" or "systemd" in make-deb.sh
  * Remove shell variable substitution from /etc/default/wolssh so it's
    compatible as a systemd EnvironmentFile
  * Fix /etc/wolssh.ini in debian/conffiles
  * Put control.tar.gz before data.tar.gz in the deb archive

+++ scripts/make-deb.sh
@@ -13,18 +13,33 @@ case $GOARCH in
     *)          ARCH=$GOARCH ;;
 esac
 
+# init system, either sysv or systemd
+: ${INIT:=systemd}
+case $INIT in
+    systemd|sysv) : ;;
+    *) echo "Invalid value for INIT ($INIT)" ; exit 1 ;;
+esac
+
 wolssh="wolssh-${GOARCH}"
 wolssh="${wolssh%-}"
 
 tmpdir="$(mktemp -d)"
-trap "rm -rf $tmpdir" EXIT
+if [[ -z $SAVETMP ]]; then
+    trap "rm -rf $tmpdir" EXIT
+else
+    echo "working in $tmpdir"
+fi
 
 echo "preparing data"
 datadir=$tmpdir/data
 install -Dm755 $wolssh $datadir/usr/bin/wolssh
-install -Dm755 debian/wolssh.init $datadir/etc/init.d/wolssh
 install -Dm644 default.ini $datadir/etc/wolssh.ini
 install -Dm644 debian/wolssh.default $datadir/etc/default/wolssh
+if [[ $INIT == sysv ]]; then
+    install -Dm755 debian/wolssh.init $datadir/etc/init.d/wolssh
+else # systemd
+    install -Dm644 debian/wolssh.service $datadir/lib/systemd/system/wolssh.service
+fi
 tar --owner=root:0 --group=root:0 -czf $tmpdir/data.tar.gz -C $datadir .
 
 echo "creating debian control"
@@ -40,4 +55,4 @@ echo '2.0' >$tmpdir/debian-binary
 pkgname="wolssh-$VERSION-$ARCH.deb"
 echo "creating $pkgname"
 rm -f $pkgname
-ar -rcs $pkgname $tmpdir/debian-binary $tmpdir/data.tar.gz $tmpdir/control.tar.gz
+ar -rcs $pkgname $tmpdir/debian-binary $tmpdir/control.tar.gz $tmpdir/data.tar.gz
